requirements.txt
amqp==5.1.0
appdirs==1.4.4
asgiref==3.5.0
astroid==2.11.2
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
billiard==3.6.4.0
black==22.3.0
celery==5.2.3
certifi==2021.10.8
click==8.1.0
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dataclasses==0.6
Deprecated==1.2.13
dill==0.3.4
Django==3.2.12
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
factory-boy==3.2.1
Faker==13.3.4
flake8==4.0.1
isort==5.10.1
kombu==5.2.4
lazy-object-proxy==1.7.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.3
pathspec==0.9.0
pip==21.2.4
platformdirs==2.5.1
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pylint==2.13.3
pyparsing==3.0.7
pyright==1.1.233
python-dateutil==2.8.2
pytz==2022.1
redis==4.2.0
regex==2022.3.15
setuptools==59.6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typed-ast==1.4.3
typing_extensions==4.1.1
vine==5.0.0
wcwidth==0.2.5
wheel==0.37.1
wrapt==1.14.0
django-celery-results==2.3.0
django-celery-beat==2.2.1
